move makeDateStringSchema to date.ts
